requirements.txt
absl-py==1.4.0
arrow==1.2.3
binaryornot==0.4.4
brotlipy==0.7.0
build==0.10.0
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.0.4
click==8.1.3
contextlib2==21.6.0
cookiecutter==2.1.1
cryptography==39.0.1
einops==0.6.1
evalutils==0.4.2
idna==3.4
imageio==2.31.1
Jinja2==3.1.2
jinja2-time==0.2.0
joblib==1.2.0
lazy_loader==0.3
MarkupSafe==2.1.3
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
ml-collections==0.1.1
networkx==3.1
numpy==1.25.0
opencv-python==4.8.0.74
packaging==21.3
pandas==2.0.2
Pillow==9.4.0
pip==23.1.2
pip-tools==6.13.0
pycparser==2.21
pyOpenSSL==23.0.0
pyparsing==3.1.0
pyproject_hooks==1.0.0
PySocks==1.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
pytorch-pretrained-vit==0.0.7
pytz==2023.3
PyWavelets==1.4.1
PyYAML==6.0
requests==2.29.0
scikit-image==0.21.0
scikit-learn==1.2.2
scipy==1.11.0
setuptools==67.8.0
SimpleITK==2.2.1
six==1.16.0
text-unidecode==1.3
threadpoolctl==3.1.0
tifffile==2023.4.12
tomli==2.0.1
torch==1.11.0
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
typing_extensions==4.6.3
tzdata==2023.3
urllib3==1.26.16
vit-pytorch==1.2.4
wheel==0.38.4
